fix(app): allow well-known files to be dropped
On my OS (Linux) files with known extensions (TXT or XML) get type set
to the actual mime-type.

This prevents dropping these files, as we previously matched for an
empty file type only (assuming that the OS does not know about BPMN,
CMMN and DMN files).

This commit improves the situation. If the OS reports a well-known type
(or a text typethat may contain a diagram), allow dragging over.